[NanaZip.Core] Update 7-Zip to 24.09. (Thanks to Igor Pavlov. Noticed by FadeMind and peashooter2.)

+++ NanaZip.Core/SevenZip/CPP/7zip/UI/Common/TempFiles.cpp
@@ -13,7 +13,8 @@ void CTempFiles::Clear()
 {
   while (!Paths.IsEmpty())
   {
-    NDir::DeleteFileAlways(Paths.Back());
+    if (NeedDeleteFiles)
+      NDir::DeleteFileAlways(Paths.Back());
     Paths.DeleteBack();
   }
 }
